% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_model_eval.R
\name{binary_model_eval}
\alias{binary_model_eval}
\title{binary_model_eval}
\usage{
binary_model_eval(modelo)
}
\arguments{
\item{modelo}{A fitted binary classification model object. Supported models include \code{glm} for logistic regression and \code{glmerMod} for mixed-effects logistic regression.}
}
\value{
A named numeric vector containing the evaluation metrics:
\item{acuracia}{Accuracy of the model.}
\item{precisao}{Precision (positive predictive value) of the model.}
\item{sensibilidade}{Sensitivity (recall) of the model.}
\item{especificidade}{Specificity of the model.}
\item{F1_Score}{F1 score of the model.}
\item{AUC}{Area under the ROC curve.}
\item{Pseudo_R2.McFadden}{McFadden's pseudo R-squared.}
\item{Pseudo_R2}{Nagelkerke's pseudo R-squared.}
\item{AIC}{Akaike Information Criterion.}
\item{BIC}{Bayesian Information Criterion.}
\item{VIF}{Variance Inflation Factor indicating multicollinearity (if applicable).}
\item{R2M}{Marginal R-squared (for GLMM).}
\item{R2c}{Conditional R-squared (for GLMM).}
\item{Status}{Indicates successful completion (1) or an error (0).}
}
\description{
Evaluates binary classification models using various metrics.
}
\details{
This function assesses the performance of binary classification models, including Generalized Linear Models (GLM)
and Generalized Linear Mixed Models (GLMM). It computes various evaluation metrics such as accuracy, precision, sensitivity,
specificity, F1 score, and area under the ROC curve (AUC). Additionally, it calculates model-specific statistics such as
pseudo R-squared values (McFadden and Nagelkerke) and information criteria (AIC, BIC). If the model is a GLMM, the function
also provides conditional and marginal R-squared values.
}
\examples{
# Example of using the binary_model_eval function with a logistic regression model
model <- glm(vs ~ am + carb, data = mtcars, family = binomial())
binary_model_eval(model)

# Example of using the function with a mixed-effects logistic regression model
#library(lmerTest)
#model_mixed <- lmerTest::glmer(as.factor(vs) ~ am + (1|carb), data = mtcars, family = binomial())
#binary_model_eval(model_mixed)

# The function returns a named vector with the calculated metrics
}
