% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table.R
\name{count_table}
\alias{count_table}
\title{count_table}
\usage{
count_table(df, var_y, var_x, sentido_percent = "col", apenas_fisher = F)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{var_y}{A string representing the name of the dependent (response) categorical variable.}

\item{var_x}{A string representing the name of the independent (explanatory) categorical variable.}

\item{sentido_percent}{A string indicating the direction in which percentages should be calculated.
Options are 'col' for column percentages or 'row' for row percentages. Default is 'col'.}

\item{apenas_fisher}{A logical value indicating whether to use only Fisher's exact test when the contingency table
has 2x2 dimensions. If \code{TRUE}, Fisher's test will always be used for 2x2 tables. Default is \code{FALSE}.}
}
\value{
A data frame containing the formatted contingency table with counts, percentages, and the results of the hypothesis test.
}
\description{
Creates a contingency table with counts, percentages, and hypothesis testing.
}
\details{
This function generates a contingency table that includes counts, row and column totals, and percentages.
Additionally, it applies appropriate hypothesis tests (Fisher's exact test or Chi-squared test) to evaluate
the association between two categorical variables. The table is formatted for clarity, making it ideal for
presentation in reports or publications.
}
\examples{
# Example of using the count_table function
data <- data.frame(treatment = c("A", "B", "A", "C", "B", "A"),
                   outcome = c("Yes", "No", "Yes", "Yes", "No", "No"))
count_table(data, "outcome", "treatment")
# Returns a contingency table with counts, percentages, and a p-value from the appropriate test
}
