% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_shapiro_test.R
\name{group_shapiro_test}
\alias{group_shapiro_test}
\title{group_shapiro_test}
\usage{
group_shapiro_test(df, col_num, col_cat, type_response = 0)
}
\arguments{
\item{df}{A data frame containing the data to be tested.}

\item{col_num}{A string representing the name of the numeric variable to be tested for normality.}

\item{col_cat}{A string representing the name of the categorical variable that defines the groups.}

\item{type_response}{An integer that specifies the type of output.
If \code{0}, the function returns a data frame with the p-values for each group.
If \code{1}, the function returns \code{TRUE} if all groups follow a normal distribution, and \code{FALSE} otherwise. Default is \code{0}.}
}
\value{
Depending on the value of \code{type_response}:
\itemize{
\item If \code{type_response = 0}: A data frame with two columns, \code{group} and \code{p_value}, showing the p-values for each group.
\item If \code{type_response = 1}: A logical value. Returns \code{TRUE} if all groups follow a normal distribution, and \code{FALSE} otherwise.
}
}
\description{
Performs the Shapiro-Wilk test for normality by group.
}
\details{
This function applies the Shapiro-Wilk test to assess the normality of a numeric variable within groups
defined by a categorical variable. The function provides an option to return either the p-values for each group
or a logical value indicating whether all groups follow a normal distribution.
}
\examples{
# Example of using the group_shapiro_test function with p-value output
data <- data.frame(response = rnorm(200, mean = 30, sd = 40),
                   desfecho = sample(c("A", "B"), size = 200, replace = TRUE, prob = c(0.6, 0.4)))
group_shapiro_test(data, 'response', 'desfecho', type_response = 0)
# Returns a data frame with p-values for each group

# Example of using the function with logical output
group_shapiro_test(data, 'response', 'desfecho', type_response = 1)
# Returns TRUE if all groups follow a normal distribution, FALSE otherwise
}
