% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fisher_applicable.R
\name{is_fisher_applicable}
\alias{is_fisher_applicable}
\title{is_fisher_applicable}
\usage{
is_fisher_applicable(df, var1, var2)
}
\arguments{
\item{df}{A data frame containing the categorical variables to be tested.}

\item{var1}{A string representing the name of the first categorical variable.}

\item{var2}{A string representing the name of the second categorical variable.}
}
\value{
A logical value. Returns \code{TRUE} if Fisher's exact test is applicable, and \code{FALSE} otherwise.
}
\description{
Determines if the Fisher's exact test is appropriate for testing the hypothesis between two categorical variables.
}
\details{
This function checks whether the Fisher's exact test is suitable for testing the association between two categorical variables.
The Fisher's test is generally more appropriate when the sample sizes are small and when the expected frequencies in any cell
of the contingency table are less than 5. This function returns \code{TRUE} if Fisher's test is recommended based on these criteria,
and \code{FALSE} otherwise.
}
\examples{
# Example of using the is_fisher_applicable function
data <- data.frame(group = c("A", "B", "A", "C", "B", "A"),
                   outcome = c("Yes", "No", "Yes", "Yes", "No", "No"))
is_fisher_applicable(data, "group", "outcome")
# Returns TRUE or FALSE based on whether Fisher's exact test is appropriate
}
