% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAStat-package.R
\docType{package}
\name{statgsa-package}
\alias{statgsa}
\alias{statgsa-package}
\title{statgsa: Statistical Analysis Tools for Generalized Linear Models}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielGabes/statgsa}
  \item Report bugs at \url{https://github.com/GabrielGabes/statgsa/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Silva dos Anjos \email{gabriel_s_anjos@yahoo.com} (\href{https://orcid.org/0009-0006-0182-904X}{ORCID})

}
\keyword{internal}
