% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_num_nonparametric_groups.R
\name{summary_num_nonparametric_groups}
\alias{summary_num_nonparametric_groups}
\title{summary_num_nonparametric_groups}
\usage{
summary_num_nonparametric_groups(df, col_num, col_cat, teste_extra = "F")
}
\arguments{
\item{df}{A data frame containing the data to be summarized.}

\item{col_num}{A string representing the name of the numeric variable to be summarized.}

\item{col_cat}{A string representing the name of the categorical variable that defines the groups.}

\item{teste_extra}{A character indicating whether to calculate the Hodges-Lehmann estimator for effect size. Set to "T" to perform this calculation; otherwise, no effect size is calculated. Default is "F".}
}
\value{
A data frame with the following columns:
\item{Variable}{The name of the numeric variable.}
\item{median \link{IQR}}{The median and interquartile range (IQR) of the numeric variable for each group and overall.}
\item{P-value}{The p-value from the statistical test (Mann-Whitney or Kruskal-Wallis).}
\item{Test_Used}{The type of statistical test applied ("Mann-Whitney" or "Kruskal-Wallis").}
\item{teste_extra}{The Hodges-Lehmann estimator and confidence interval, if calculated.}
}
\description{
Generates a summary table of median and interquartile range for a numeric variable across groups, with optional Hodges-Lehmann estimator.
}
\details{
This function calculates the median and interquartile range (IQR) of a numeric variable within groups defined by a categorical variable.
It also performs non-parametric tests (Mann-Whitney for two groups or Kruskal-Wallis for more than two groups) and optionally calculates the Hodges-Lehmann estimator for pairwise comparisons.
}
\examples{
# Example of using the summary_num_nonparametric_groups function
data <- data.frame(value = c(1.2, 2.3, 3.1, 4.7, 5.0, 2.1, 3.2, 4.8, 5.6, 6.7),
                   group = c("A", "B", "A", "B", "A", "A", "B", "B", "A", "B"))
summary_num_nonparametric_groups(data, 'value', 'group', teste_extra = "T")
# Returns a summary table with median, IQR, p-value, test used, and Hodges-Lehmann estimator.
}
