% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_num_parametric.R
\name{summary_num_parametric}
\alias{summary_num_parametric}
\title{summary_num_parametric}
\usage{
summary_num_parametric(df, col_num, digits = 2)
}
\arguments{
\item{df}{A data frame containing the data to be summarized.}

\item{col_num}{A string representing the name of the numeric variable to be summarized.}

\item{digits}{An integer indicating the number of decimal places to be used in the formatted output.}
}
\value{
A data frame with the following columns:
\item{Variable}{The name of the numeric variable.}
\item{mean}{The mean of the numeric variable.}
\item{std}{The standard deviation of the numeric variable.}
}
\description{
Generates a summary table with the mean and standard deviation for a numeric variable.
}
\details{
This function calculates the mean and standard deviation of a numeric variable from a data frame,
providing a summary that is useful for parametric analyses. The results are returned in a formatted table
with the variable name included for clarity.
}
\examples{
# Example of using the summary_num_parametric function
data <- data.frame(value = c(1.2, 2.3, 3.1, 4.7, 5.0))
summary_num_parametric(data, 'value')
# Returns a data frame with the mean and standard deviation of 'value'
}
