% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_model_analysis.R
\name{univariate_model_analysis}
\alias{univariate_model_analysis}
\title{Univariate Model Analysis}
\usage{
univariate_model_analysis(modelo)
}
\arguments{
\item{modelo}{A model object of class \code{glm}. The model should be either logistic regression (with family \code{binomial}) for Odds Ratio calculation or quasibinomial regression for Relative Risk calculation.}
}
\value{
A data frame containing the calculated Odds Ratio or Relative Risk, along with their confidence intervals and p-values. If the model is not supported, the function returns \code{NULL}.
}
\description{
This function performs a univariate analysis on a given model, calculating either the Odds Ratio (OR) or Relative Risk (RR) depending on the type of model provided.
The function handles models of class \code{glm} with either a \code{binomial} or \code{quasibinomial} family.
}
\examples{
# Logistic regression example
model <- glm(vs ~ am + carb, data = mtcars, family = binomial())
univariate_model_analysis(model)

# Quasibinomial regression example
model <- glm(vs ~ am + carb, data = mtcars, family = poisson())
univariate_model_analysis(model)

}
